Move flattenItem and unflattenItem to a common location
