--- src/webextension/list/common.js
@@ -1,7 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-// This is a string instead of a symbol since we need to be able to use it as a
-// React `key` inside a list.
-export const NEW_ITEM_ID = "new-item";
